gdb/python: add TargetConnection.send_remote_packet method
This commit adds a new method to the gdb.TargetConnection object type:
'send_remote_packet'.  This new method is equivalent to the 'maint
packet' CLI command, it allows a custom packet to be sent to a remote
target.

Not all gdb.TargetConnection instances will support
send_remote_packet, e.g. the 'native' target doesn't.  If the
send_remote_packet method is invoked on a TargetConnection that is not
'remote' or 'extended-remote' then an exception is thrown.

The result of calling TargetConnection.send_remote_packet is a string
containing the reply that came from the remote.