gdb/python: implement the print_insn extension language hook
This commit extends the Python API to include disassembler support,
and additionally provides a syntax highlighting disassembler.

The motivation for this commit was to provide an API by which the user
could write Python scripts that would augment the output of the
disassembler.

To achieve this I have followed the model of the existing libopcodes
disassembler, that is, instructions are disassembled one by one.  This
does restrict the type of things that it is possible to do from a
Python script, i.e. all additional output has to fit on a single line,
but this was all I needed, and creating something more complex would,
I think, require greater changes to how GDB's internal disassembler
operates.

It was only once I had a working prototype that I realised I could
very easily use this to perform syntax highlighting on GDB's
disassembly output, so I've included that too.  The new commands added
are:

  set style disassembly on|off
  show style disassembly

which enable or disable disassembly syntax highlighting.

The disassembler API is contained in the new gdb.disassembler module,
which defines the following classes:

  DisassembleInfo

      Similar to libopcodes disassemble_info structure, has read-only
  attributes: address, string, length, architecture, and
  can_emit_style_escape.  And has methods: read_memory, set_result,
  and memory_error.

      Each time GDB wants an instruction disassembled, an instance of
  this class is passed to a user written disassembler, by reading the
  attributes, and calling the methods, the user can perform
  disassembly, and set the result within the DisassembleInfo instance.

  Disassembler

      This is a base-class which user written disassemblers should
  inherit from, just provides base implementations of __init__ and
  __call__ which the user written disassembler should override.

The gdb.disassembler module also provides the following functions:

  register_disassembler

      This function registers an instance of a Disassembler sub-class
  as a disassembler, either for one specific architecture, or, as a
  global disassembler for all architectures.

  format_address

      This wraps GDB's print_address function, converting an address
  into a string that can be placed into disassembler output.

  syntax_highlight

      This adds syntax highlighting escapes to some disassembler
  output, users can call this from their own custom disassemblers to
  retain syntax highlighting, this function handles switching syntax
  highlighting off, or the case where the pygments library is not
  available.

  builtin_disassemble

      This provides access to GDB's builtin disassembler.  A common
  user case that I see is augmenting the existing disassembler
  output.  The user code can call this function to have GDB
  disassemble the instruction in the normal way, and then the user can
  tweak the output before returning that as the result.  This function
  also provides a mechanism to intercept the disassemblers reads of
  memory, thus the user can adjust what GDB sees when it is
  disassembling.

The included documentation provides a more detailed description of the
API.