[wip] gdb/doc: document the buffalo specific extensions to gdb
This is a work in progress start on documenting the buffalo specific
GDB extensions.

The only extensions so far relate to the custom instruction
disassembler support.

I've documented the two new GDB settings:

  set/show custom-instruction-display
  set/show custom-instruction-filename

[TODO]: I've started a section in which the format of the custom instruction
file should be described, though this section needs to be filled in.

I've also had a pass at attempting to describe how the remote protocol
handling needs to be extended to support fetching the instruction XML
file from a remote, though I'm worried that this section might not be
very clear.

[TODO]: Within this document I have also used the project name
"buffalo", rather than using the clients name specifically.  I don't
know if this needs to be changed, I'll leave that for others to
decide.