gdb/custominsn: Add $dest support, and merge some common code
Add a new Insn.expand_format_string method that takes a dictionary,
and expands the format_string based on the key/value pairs in the
dictionary.

The Insn.expand_format_string method also adds support for the '$dest'
field, which is filled in based on the instruction's address and the
'imm' field value, this can be useful for branch/jump instructions.

Finally, all of the gen_instr_assembly methods now make use of
expand_format_string.